requirements.txt
psutil==5.4.8
PyCRC==1.21
pyserial==3.4
